+++ src/index.ts
@@ -1,3 +1,4 @@
 export { default as Account } from "./account";
 export { getTangleApi } from "./api";
 export { createProfile, deleteProfile, updateProfile } from "./roles";
+export * from "./types";
